datasetiloc fix
looked inside of csv goes up to 43 columns, 42 columns of data and column 43 is outcome win/loss/draw

changed x and y to fit the csv columns

now all that needs to be done is the data needs to be encoded. Will read the python documentation to try and find the most efficient way to do this tomorrow or on Monday.